+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ x.x.x (x-x-x)
 -------------------
 
 * Added tests for environment variables and deployments commands.
+* Added support to browse organisations.
 
 3.17.2 (2023-03-16)
 -------------------